pyproject.toml
[project]
name = "ga-cli"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Alan Wu", email = "lunw1024@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "mcp>=1.10.1",
    "python-dotenv>=1.1.1",
    "rich>=14.0.0",
    "tomlkit>=0.13.3",
    "typer>=0.16.0",
]

[project.scripts]
ga = "ga_cli.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.4.1",
]
